move all scripts into ts files
