+++ src/components/ThemeButton.astro
@@ -23,5 +23,5 @@ import { Icon } from "astro-icon/components";
 </style>
 
 <script>
-  import "@/scripts/theme.js";
+  import "@/scripts/theme";
 </script>
