+++ src/pages/index.astro
@@ -373,5 +373,5 @@ const sortedPosts = sortPostsByDate(allPosts, 3);
 </style>
 
 <script>
-  import "../scripts/observer.js";
+  import "@/scripts/observer";
 </script>
