+++ src/scripts/navbar.ts
